Updates befor the first full crop run.
